requirements.txt
numpy==2.1.3
opencv-python==4.10.0.84
pillow==11.0.0
pillow_heif==0.20.0
PyExifTool==0.5.6
numpy==2.1.3
